backend/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.2.0
Django==4.2.3
django-cors-headers==4.2.0
django-crontab==0.7.1
djangorestframework==3.14.0
idna==3.4
Pillow==10.0.0
psycopg2-binary==2.9.6
pyotp==2.9.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.3
